primary_requirements.txt
face_recognition_models
Click
dlib
numpy
Pillow
scipy